More work on the frequency display panel it kinda worked now it doesn't dunno why, see run() method of FrequencyDisplayPanel

+++ src/audio/files/AudioFile.java
@@ -11,5 +11,5 @@ public interface AudioFile {
     public int[] getAllSamples();
     public int getNumberOfSamples();
     
-    public int[] getChunk(int chunk_size);
+    public double[] getChunk(int chunk_size);
 }
